app: Break validation up into a module
