+++ tracecat/dsl/validation.py
@@ -9,7 +9,7 @@ from tracecat.dsl.common import DSLInput
 from tracecat.dsl.models import TriggerInputs
 from tracecat.expressions.expectations import ExpectedField, create_expectation_model
 from tracecat.logger import logger
-from tracecat.types.validation import ValidationResult
+from tracecat.validation.models import ValidationResult
 
 LIST_PATTERN = re.compile(r"list\[(?P<inner>(\$)?[a-zA-Z]+)\]")
