+++ tracecat/validation/models.py
@@ -1,6 +1,5 @@
 import json
 from collections.abc import Mapping
-from datetime import datetime, timedelta
 from typing import Any, Literal
 
 from pydantic import BaseModel
@@ -35,14 +34,3 @@ class ExprValidationResult(ValidationResult):
 
 class SecretValidationResult(ValidationResult):
     """Result of validating credentials."""
-
-
-VALIDATION_TYPES = {
-    "duration": timedelta,
-    "datetime": datetime,
-    "str": str,
-    "int": int,
-    "float": float,
-    "bool": bool,
-    "any": Any,
-}
